Added faceUp variable
I created a third card object as card3 to represent the deck. I added a faceUp instance variable in the __init__ method of cards.py and set it to False. In the newDeal method of project 9 I added two lines that change the faceUp variable of card1 and card2 to True in order to simulate "flipping" the card over. I added a fileName4 to the refresh images method that is always "b.gif". I added if/else statements for each card to determine if they are face up or not and display their face or the card back.